Refactor(Players Profile): Refactor player profile APIs to be RESTful
In this commit, my focus was on refactoring the player profile APIs to better align with RESTful conventions. As part of this, I also had to modify parts of the framework. For example, commands needed to be able to return results, and for mapping objects to API responses, I had to integrate the AutoMapper library into the framework.