Refactor(Structure): Reorganized solution and separated service layers
In this commit, I performed a structural reorganization of the projects and solution, leading to the separation of the service layers into a dedicated folder. Each service now has its own src and test directories; currently, only the src folder is populated. Additionally, I created a contracts project for the game service, so that the background projects now reference this project instead of the domain.