Merge develop branch into master (#34)
* refactore: change TestAuthHandler based on ISystemClock is obsolete breacking change in dotnet 8

* feat: Api versioning for players bounded context implemented.

* refactor: Players context removed and Framework.ApplicationServices renamed to Framework.Application

* refctor: Event strore persistence tools added to framework and player aggregate ragistration method implemented.

* feature: player repository implemented.

* feature: player registration command and command handler implemented.

* project structure changed.

* event store client updated.

* event store bugs fixed and swagger identity server configuration added.

* Query model for player and QueryDbContext created.

* sql projection added.

* find cursour position

* sql server projection completed.

* view profile api implemented.

* Refactor(Players Profile): Refactor player profile APIs to be RESTful

In this commit, my focus was on refactoring the player profile APIs to better align with RESTful conventions. As part of this, I also had to modify parts of the framework. For example, commands needed to be able to return results, and for mapping objects to API responses, I had to integrate the AutoMapper library into the framework.

* Refactor(Project Structure): Consolidated framework libraries into a single directory

In this commit, I restructured the project by consolidating all framework libraries into a single directory. Previously, these libraries were scattered across different subfolders under the framework, which led to build errors. I moved all the framework libraries to the "framework" folder, placing them at the same level. Additionally, I resolved the build issues in projects and libraries that depended on the framework due to this restructuring.

* Refactor(Framework): Simplify framework layers and consolidate related tasks

In this commit, I aimed to reduce the layers related to the framework section by consolidating related tasks into the libraries, making the framework layers fewer. Next, I focused on the projection code in the background section, moving common functionalities that could be reused in future projections into the framework. Finally, I renamed the project related to SQL projection.

* Refactor(IDP): Simplify IDP project by removing unnecessary layers

In this commit, I completely refactored the IDP project. I felt that there were too many additional layers that were not needed for this project, so I removed them to simplify the structure.

* Refactor(Structure): Reorganized solution and separated service layers

In this commit, I performed a structural reorganization of the projects and solution, leading to the separation of the service layers into a dedicated folder. Each service now has its own src and test directories; currently, only the src folder is populated. Additionally, I created a contracts project for the game service, so that the background projects now reference this project instead of the domain.

* Feat(Player Registration): Prevent duplicate user registrations and standardize error responses

In this commit, I implemented a domain service to prevent duplicate user registrations in the player profile. Additionally, I created a consistent error response structure for API responses. This involved developing a middleware that returns a uniform error output for all requests encountering errors.

* Refactor: Fix compiler warnings

In this commit, I focused on resolving various compiler warnings. This helps to clean up the codebase and ensure a smoother development process.